Simple WiC64 telnet client compatible with firmware 2.0.0 and later.
This client is intended to replace the legacy "Telnet Demo", as that
program no longer works with firmware 2.0.0 and later due to necessary
improvements of the TCP READ command.

To be specific, the legacy firmware simply polls the connection for 500ms
on each command invocation, reading data from the connection byte by byte
during this timeframe. So regardless of whether there were 10, 1000 or
even no bytes bytes available at all, the command always took 500ms,
making the client slow to send keystrokes and to receive and display
the keystrokes echoed back from the server.

This program improves performance in this regard. If the server is
reasonably fast, the roundtrip delay is barely noticeable. Nevertheless
this client also allows the user to enter up to 80 charactes on the c64
and to send them in bulk, as servers may always be slow to respond.

In addition, this client properly handles timeout and error conditions,
including temporary loss of WiFi connectivity. In case of timeouts or
tcp errors, the user is given the option of either trying to continue or
or to abort the current session. In case of connection loss, reconnection
to the current server is attempted on retry.

Like the original version, this program is still somewhat simple, lacking
proper scrollback buffer support and as well as a decent input editor.
It also does not handle telnet IAC codes at this time. For now, it is
mainly intended  as a drop in replacement for the legacy program.