Changed retry/abort keys to F3 and F1, removed "Connecting" message
Changed keybindings for the retry and abort options to F3 and F1.

I was writing a message when I ran into a timeout and was just about to
type "read", so I did a (r)etry and (a)bort before I knew it.
Function keys are less likely to be hit accidentally.

Removed "Connect <host:port>..." because it just flashes by anyways.